pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Yossef-Dawoad <yossefdawoad15@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
alembic = "^1.12.1"
sqlalchemy = "^2.0.23"
uvicorn = "^0.24.0.post1"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0.post1"
python-multipart = "^0.0.6"
passlib = {extras = ["argon2"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.6"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
extend-select = [
    'C4',
    "F",   # pyflakes
    "E", # Errors
    "W", # warnnings
    "RUF", # Ruff-specific rules
    "PERF", #PREFlint
    "PL", # PYLINT
    "UP",  # pyupgrade
    'SIM', #sim for code simplification
    'TCH', #tch for type checking
    'F707', # Syntax error in forward annotation
    'I001', # Import block is un-sorted or un-formatted
    'I002', # Missing required import:
    'COM', # FLAKE COMMA
    'ANN', # flake8-annotations
    'PYI052', #Need type annotation for
    'DTZ', #FLAKE datetime
    'ASYNC', #FLAKE ASYNC
    'PLE1700'
]
show-fixes = true
target-version = "py310"
line-length = 88
ignore = ['E501', 'SIM115', 'ANN101', "UP007", "E701"]
extend-exclude = ['alembic']
